# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "description": "List notifications for the current authenticated user",
        "operationId": "NotificationsController_getNotifications",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Optional filter by notification category, 0 for User, 1 for System, 2 for Trading",
            "schema": {
              "type": "number",
              "enum": [
                0,
                1,
                2
              ]
            }
          },
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "description": "Optional filter by read status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get unread notification count",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/read-all": {
      "patch": {
        "description": "Mark all notifications as read for current user",
        "operationId": "NotificationsController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Read all notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{notificationId}/read": {
      "patch": {
        "description": "Mark a notification as read for current user",
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "notificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Read notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/admin/transaction": {
      "get": {
        "operationId": "AdminPaymentController_getTransactions",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Transaction type",
            "schema": {
              "type": "string",
              "enum": [
                "DEPOSIT",
                "WITHDRAW",
                "TRANSFER",
                "TRADE_BUY",
                "TRADE_SELL",
                "DIVIDEND",
                "FEE"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Transaction status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "COMPLETED",
                "FAILED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date (ISO format)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date (ISO format)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Admin Get Transaction Histories",
        "tags": [
          "Admin - Payment"
        ]
      }
    },
    "/admin/withdraw/{transactionId}/status": {
      "patch": {
        "operationId": "AdminPaymentController_updateWithdrawStatus",
        "parameters": [
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateWithdrawStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Admin Update Withdraw Transaction Status",
        "tags": [
          "Admin - Payment"
        ]
      }
    },
    "/admin/sweep-funds": {
      "post": {
        "operationId": "AdminPaymentController_sweepFunds",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSweepFundsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Sweep Funds (Gom quỹ từ ví người dùng về ví tổng)",
        "tags": [
          "Admin - Payment"
        ]
      }
    },
    "/orders/bulk-cancel": {
      "post": {
        "description": "Cancel multiple open orders at once.",
        "operationId": "OrdersController_bulkCancel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCancelOrdersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Bulk cancel orders",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{orderId}": {
      "patch": {
        "description": "Update mutable fields or cancel by setting cancel=true",
        "operationId": "OrdersController_updateOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update or cancel order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders": {
      "get": {
        "description": "List orders with filters and sorting",
        "operationId": "OrdersController_getOrders",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "side",
            "required": false,
            "in": "query",
            "description": "Order side",
            "schema": {
              "type": "string",
              "enum": [
                "BUY",
                "SELL"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Order status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "OPEN",
                "PARTIALLY_FILLED",
                "FILLED",
                "CANCELLED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "assetId",
            "required": false,
            "in": "query",
            "description": "Asset ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List orders",
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "description": "Create a new trading order (MARKET vs PENDING)",
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create order",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "RWA Backend API",
    "description": "API documentation for the RWA graduation project - Real World Assets Platform",
    "version": "1.0.0",
    "contact": {
      "name": "RWA Team",
      "url": "",
      "email": "support@rwa.com"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      }
    },
    "schemas": {
...
      "AdminUpdateWithdrawStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New status",
            "enum": [
              "COMPLETED",
              "FAILED",
              "CANCELLED"
            ]
          },
          "transactionHash": {
            "type": "string",
            "description": "Required if status is COMPLETED"
          },
          "rejectedReason": {
            "type": "string",
            "description": "Required if status is FAILED or CANCELLED"
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid",
            "description": "Idempotency key to prevent double execution"
          }
        },
        "required": [
          "status",
          "idempotencyKey"
        ]
      },
      "AdminSweepFundsDto": {
        "type": "object",
        "properties": {
          "chainId": {
            "type": "string",
            "enum": [
              "1",
              "56",
              "97",
              "11155111"
            ],
            "description": "Chain ID"
          },
          "destinationWallet": {
            "type": "string",
            "description": "Destination wallet address to sweep funds to"
          },
          "privateKey": {
            "type": "string",
            "description": "Private key of the admin wallet used for sweeping"
          }
        },
        "required": [
          "chainId",
          "destinationWallet",
          "privateKey"
        ]
      },
      "BulkCancelOrdersDto": {
        "type": "object",
        "properties": {
          "orderIds": {
            "minItems": 1,
            "description": "Array of order IDs to cancel",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "orderIds"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "cancel": {
            "type": "boolean",
            "description": "Set to true to cancel the order. If true, other fields are ignored"
          },
          "price": {
            "type": "object",
            "description": "New limit/stop price. Only applicable for non-MARKET orders"
          },
          "stopLossPrice": {
            "type": "object",
            "description": "New stop loss price. Optional, set to null to remove"
          },
          "takeProfitPrice": {
            "type": "object",
            "description": "New take profit price. Optional, set to null to remove"
          }
        }
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "description": "Asset ID"
          },
          "side": {
            "type": "string",
            "description": "Order side",
            "enum": [
              "BUY",
              "SELL"
            ]
          },
          "type": {
            "type": "string",
            "description": "Order type",
            "enum": [
              "MARKET",
              "LIMIT"
            ]
          },
          "quantity": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d+)?$",
            "description": "Order quantity"
          },
          "price": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d+)?$",
            "description": "Limit price (required for LIMIT orders, ignored for MARKET)"
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid",
            "description": "Client-generated UUID to prevent duplicate submissions"
          }
        },
        "required": [
          "assetId",
          "side",
          "type",
          "quantity"
        ]
      },
...
      "ValidationErrorItem": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Machine-friendly field-level error code."
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "code",
          "t"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string",
            "enum": [
              "internal-error",
              "not-found",
              "bad-request",
              "validation-error",
              "rate-limit-exceeded",
              "duplicate-request",
              "unauthorized",
              "forbidden",
              "token-expired",
              "asset-not-found",
              "invalid-token",
              "token-not-found",
              "session-not-found",
              "session-revoked",
              "session-expired",
              "user-not-found",
              "user-not-active",
              "user-already-exists",
              "invalid-credentials",
              "reference-code-not-found",
              "reference-code-invalid",
              "reference-code-already-set",
              "password-attempt-limit-exceeded",
              "otp-not-found",
              "invalid-otp",
              "invalid-otp-token",
              "mfa-already-enabled",
              "mfa-not-enabled",
              "invalid-mfa-code",
              "challenge-required",
              "challenge-not-found",
              "challenge-expired",
              "challenge-already-used",
              "challenge-invalid-factor",
              "challenge-payload-mismatch",
              "account-not-found",
              "account-not-active",
              "position-not-found",
              "position-already-closed",
              "feedback-not-found",
              "post-not-found",
              "post-unauthorized",
              "comment-not-found",
              "comment-unauthorized",
              "support-ticket-not-found",
              "support-unauthorized",
              "insufficient-balance",
              "transfer-to-same-account",
              "transfer-not-allowed-with-open-position",
              "external-api-error",
              "invalid-quantity",
              "price-not-found"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error summary for clients/logging."
          },
          "t": {
            "type": "string",
            "format": "date-time"
          },
          "errors": {
            "description": "Detailed error information by field, domain rule, or external provider.",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorItem"
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "traceId": {
            "type": "string",
            "description": "Request trace id for debugging across services."
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request payload or failed domain validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is missing, invalid, revoked, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed to access or mutate this resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded or too many OTP/password attempts.",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Unexpected server error or upstream dependency failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}